Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
pydantic
pydantic[email]
python-dotenv
python-multipart